refactored signUp validator
