+++ app/helpers/users/generate_password.js
